fix: make sure to unmount on close
