+++ src/Molecules/Modal/Modal.stories.tsx
@@ -115,47 +115,36 @@ storiesOf('Molecules | Modal', module)
   })
   .add('Uncontrolled behavior', () => {
     const Example = () => {
-      const [open, setOpen] = useState(false);
-
-      const onOpen = () => {
-        setOpen(true);
-      };
-
       return (
         <div>
-          <button type="button" onClick={onOpen}>
-            Open modal
-          </button>
           <ScrollMaker />
-          {open && (
-            <Modal title="Dialog information">
-              <Box mb={2}>
-                <Typography type="primary" as="p">
-                  Modals should be used with care as they are quite intrusive on the user experience
-                  and demand immediate attention (while also blocking all other actions on the
-                  site). Always consider if you can solve a problem in another way first before you
-                  choose to go with the modal.
-                </Typography>
-              </Box>
-              <Box mb={2}>
-                <Typography type="primary" as="p">
-                  That being said they are a good tool if you need to grab the users attention,
-                  either to communicate something very important or make them take an action before
-                  proceeding.
-                </Typography>
-              </Box>
-              <Box mb={2}>
-                <Typography type="primary" as="p">
-                  Nielsen/Norman has an excellent article about their usage here
-                </Typography>
-              </Box>
+          <Modal title="Dialog information">
+            <Box mb={2}>
               <Typography type="primary" as="p">
-                <a href="https:// www.nngroup.com/articles/modal-nonmodal-dialog/">
-                  https:// www.nngroup.com/articles/modal-nonmodal-dialog/
-                </a>
+                Modals should be used with care as they are quite intrusive on the user experience
+                and demand immediate attention (while also blocking all other actions on the site).
+                Always consider if you can solve a problem in another way first before you choose to
+                go with the modal.
               </Typography>
-            </Modal>
-          )}
+            </Box>
+            <Box mb={2}>
+              <Typography type="primary" as="p">
+                That being said they are a good tool if you need to grab the users attention, either
+                to communicate something very important or make them take an action before
+                proceeding.
+              </Typography>
+            </Box>
+            <Box mb={2}>
+              <Typography type="primary" as="p">
+                Nielsen/Norman has an excellent article about their usage here
+              </Typography>
+            </Box>
+            <Typography type="primary" as="p">
+              <a href="https:// www.nngroup.com/articles/modal-nonmodal-dialog/">
+                https:// www.nngroup.com/articles/modal-nonmodal-dialog/
+              </a>
+            </Typography>
+          </Modal>
         </div>
       );
     };
