+++ src/Molecules/Modal/Modal.types.ts
@@ -2,9 +2,12 @@ export type Props = {
   className?: string;
   onClose?: Function;
   title?: string;
-  open?: boolean;
 };
 
 export type DialogProps = {
   show: boolean;
 };
+
+export type WrapperProps = {
+  open?: boolean;
+} & Props;
