+++ src/common/utils.ts
@@ -42,6 +42,7 @@ export const isUndefined = (x: any): x is undefined => typeof x === 'undefined';
 export const isElement = (x: any): x is React.ReactNode => React.isValidElement(x);
 export const isNumber = (x: any): x is number => typeof x === 'number';
 export const isBoolean = (x: any): x is boolean => typeof x === 'boolean';
+export const isFunction = (x: any): x is Function => typeof x === 'function';
 
 export const pickAriaAttributes = R.pickBy((_, key: string) => R.test(/^aria-/, key));
